fix: address latest shortcuts review comments
